partial rollback of a previous development of LagBFunction tester
This tester includes convex/concave optimization and bounds on the variables, but *not* flow bounds since these were giving problems I was not able to identify. This version passes all tests with transportation problems either easy or not, provided they are not changed. Hence it is used as a stable base for development of the full version with bounds on the flow.