Minor changes
Added a BPar.txt file for the tests of MMCFBlock in order to decide the type of relaxation and if use or not the strong forcing constraints.
Corrected also the BPar file for the test LagrangianDualSolver_MMCF in order to be coerent with the new encoding of relaxation types.
Added a space in the print of dual variables for LagrangianDualSolver_MMCF .